refactor: Table Widget Entry-Point Responsibilities
